requirements.txt
torch~=2.5.0
torchvision~=0.20.0
streamlit~=1.39.0
numpy~=1.26.4
pillow~=10.4.0